pyproject.toml
[tool.poetry]
name = "goose-checker"
version = "0.1.0"
description = ""
authors = ["apaslaw <adam.paslawski@morningstar.com>"]
readme = "README.md"
packages = [{include = "goose_checker"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.14"
langchain-openai = "^0.1.1"
pydantic = "^2.6.4"
tiktoken = "^0.6.0"
gitpython = "^3.1.43"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
goose-checker = "goose_checker.cli:cli_main"
